WIP: operator impersonation of admin-robot service account
Without explicit dependency declaration on IAM binding, a singular run is likely to fail at the very end when attempting to acquire an access token. This is because the action has a race condition with the creation of IAM policy on the service account enabling access token creation. This is illustrated by shenanigans-5.dot