Example deployment: GKE node service account
It appears that IAM Conditions are limited to only a subset of supported resources. So we assign a Project IAM Admin role to admin-robot service account and now it can manage IAM policies for the node service accounts that it creates. One such service account is created as an example with the minimal viable set of roles assigned so that it can be used as an identity for a GKE cluster node.